rename crate-scope macro file to not shadow re-export
