+++ bindings/wasm/identity_wasm/src/helper_macros.rs
